+++ services/applicants-service/src/main/java/ru/volzhanin/applicantsservice/repository/RegionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.volzhanin.applicantsservice.entity.Region;
 
 @Repository
-public interface RegionRepository extends JpaRepository<Region, Integer> {
+public interface RegionRepository extends JpaRepository<Region, Long> {
 }
